Fixed bug where only 1 vbo was drawn.

+++ src/main.c
@@ -12,7 +12,7 @@ int main() {
     VOX__play(&error);
 
     // print error code
-    printf("Error Code: %llu\nError Log: \n%s\n", error.p_type, error.p_opengl_log);
+    printf("Error Code: %llu\nError Log: \n%s\n", (unsigned long long)error.p_type, error.p_opengl_log);
 
     // tell player game is done
     printf("Thanks for playing!\n");
